Removed TablePreview component in favor of TableWidget
