+++ frontend/src/components/WidgetRender.tsx
@@ -7,6 +7,7 @@ import {
   ImageWidget,
   MetricsWidget,
 } from "../models";
+import { useWidgetDataset } from "../hooks";
 import ChartWidgetComponent from "../components/ChartWidget";
 import TableWidgetComponent from "../components/TableWidget";
 import TextWidget from "../components/TextWidget";
@@ -17,15 +18,26 @@ interface Props {
   widget: Widget;
 }
 
-function WidgetRender(props: Props) {
-  const widget = props.widget;
+function WidgetRender({ widget }: Props) {
+  const { json, jsonHeaders } = useWidgetDataset(widget);
+
   switch (widget.widgetType) {
     case WidgetType.Text:
       return <TextWidget widget={widget} />;
     case WidgetType.Chart:
       return <ChartWidgetComponent widget={widget as ChartWidget} />;
     case WidgetType.Table:
-      return <TableWidgetComponent widget={widget as TableWidget} />;
+      const tableWidget = widget as TableWidget;
+      return (
+        <TableWidgetComponent
+          title={tableWidget.showTitle ? tableWidget.content.title : ""}
+          summary={tableWidget.content.summary}
+          headers={jsonHeaders}
+          data={json}
+          summaryBelow={tableWidget.content.summaryBelow}
+        />
+      );
+
     case WidgetType.Image:
       return <ImageWidgetComponent widget={widget as ImageWidget} />;
     case WidgetType.Metrics:
